+++ app/iam-shared-kernel/src/main/java/com/chinda/iam_shared_kernel/model/Platform.java
@@ -1,0 +1,6 @@
+package com.chinda.iam_shared_kernel.model;
+
+public enum Platform {
+    GOOGLE(),
+    KAKAO();
+}
