--- app/iam-shared-kernel/src/main/java/com/chinda/user_shared_kernel/model/Gender.java
@@ -1,15 +1,0 @@
-package com.chinda.user_shared_kernel.model;
-
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-public enum Gender {
-    MALE("male", "남자"),
-    FEMALE("female", "여자");
-
-    private final String engValue;
-    private final String korValue;
-}
