--- app/iam-shared-kernel/src/main/java/com/chinda/user_shared_kernel/model/Platform.java
@@ -1,6 +1,0 @@
-package com.chinda.user_shared_kernel.model;
-
-public enum Platform {
-    GOOGLE(),
-    KAKAO();
-}
