+++ app/iam/src/main/java/com/chinda/iam/application/IdentityService.java
@@ -1,0 +1,11 @@
+package com.chinda.iam.application;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class IdentityService {
+
+
+}
