+++ app/iam/src/main/java/com/chinda/iam/application/exceptions/UserNotRegisteredException.java
@@ -1,0 +1,4 @@
+package com.chinda.iam.application.exceptions;
+
+public final class UserNotRegisteredException extends RuntimeException{
+}
