+++ app/iam/src/main/java/com/chinda/iam/domain/access/dto/KakaoUserResponse.java
@@ -1,0 +1,22 @@
+package com.chinda.iam.domain.access.dto;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import jakarta.validation.constraints.NotNull;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class KakaoUserResponse implements OAuthUserResponse {
+
+    @NotNull
+    private Long id;
+
+    @Override
+    public Long getSocialId() {
+        return this.id;
+    }
+}
