+++ app/iam/src/main/java/com/chinda/iam/domain/access/dto/OAuthTokenResponse.java
@@ -1,0 +1,5 @@
+package com.chinda.iam.domain.access.dto;
+
+public interface OAuthTokenResponse {
+    public String getAccessToken();
+}
