+++ app/iam/src/main/java/com/chinda/iam/domain/access/dto/OAuthUserResponse.java
@@ -1,0 +1,5 @@
+package com.chinda.iam.domain.access.dto;
+
+public interface OAuthUserResponse {
+    public Long getSocialId();
+}
