--- app/iam/src/main/java/com/chinda/user/UserApplication.java
@@ -1,20 +1,0 @@
-package com.chinda.user;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.ImportAutoConfiguration;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.domain.EntityScan;
-import org.springframework.cloud.openfeign.EnableFeignClients;
-import org.springframework.cloud.openfeign.FeignAutoConfiguration;
-
-@SpringBootApplication
-@EnableFeignClients("com.chinda.user")
-@ImportAutoConfiguration({FeignAutoConfiguration.class})
-@EntityScan("com.chinda.user_shared_kernel.*")
-public class UserApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(UserApplication.class, args);
-	}
-
-}
