--- app/iam/src/main/java/com/chinda/user/application/IdentityService.java
@@ -1,11 +1,0 @@
-package com.chinda.user.application;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class IdentityService {
-
-
-}
