--- app/iam/src/main/java/com/chinda/user/application/exceptions/UserNotRegisteredException.java
@@ -1,4 +1,0 @@
-package com.chinda.user.application.exceptions;
-
-public final class UserNotRegisteredException extends RuntimeException{
-}
