--- app/iam/src/main/java/com/chinda/user/domain/access/OAuthAgreedUserFactory.java
@@ -1,6 +1,0 @@
-package com.chinda.user.domain.access;
-
-public interface OAuthAgreedUserFactory {
-    OAuthAgreedUser getOAuthAgreedUser(String authCode);
-
-}
