--- app/iam/src/main/java/com/chinda/user/domain/access/dto/OAuthUserResponse.java
@@ -1,5 +1,0 @@
-package com.chinda.user.domain.access.dto;
-
-public interface OAuthUserResponse {
-    public Long getSocialId();
-}
