--- app/iam/src/main/java/com/chinda/user/web/iam/dto/request/LoginVo.java
@@ -1,20 +1,0 @@
-package com.chinda.user.web.iam.dto.request;
-
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import jakarta.validation.constraints.NotBlank;
-
-@Getter
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-@AllArgsConstructor(access = AccessLevel.PRIVATE)
-public class LoginVo {
-
-    @NotBlank(message = "플랫폼명은 필수 값입니다.")
-    private String platformName;
-    @NotBlank(message = "사용자 코드는 필수 값입니다.")
-    private String authCode;
-
-}
