+++ app/iam/src/test/java/com/chinda/user/IAMApplicationTests.java
@@ -1,0 +1,13 @@
+package com.chinda.user;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class IAMApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
