--- app/iam/src/test/java/com/chinda/user/UserApplicationTests.java
@@ -1,13 +1,0 @@
-package com.chinda.user;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class UserApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
